+++ tests/testthat/test-r-processing.R
@@ -13,14 +13,15 @@ test_that("R file processing works and creates vignettes", {
   expect_equal(
     basename(
       package_build(
-        file.path(tempdir(), "rfiletest")
+        file.path(tempdir(), "rfiletest"), 
+        install = TRUE
       )
     ),
     "rfiletest_1.0.tar.gz"
   )
   v <- vignette(package = "rfiletest")
   expect_equal(v$results[, "Item"], "rfileTest")
-
+  
   unlink(file.path(tempdir(), "rfiletest"),
     recursive = TRUE,
     force = TRUE
@@ -39,7 +40,8 @@ test_that("R file processing works and creates vignettes", {
   expect_equal(
     basename(
       package_build(
-        file.path(tempdir(), "rfiletest")
+        file.path(tempdir(), "rfiletest"),
+        install = TRUE
       )
     ),
     "rfiletest_1.0.tar.gz"
@@ -50,4 +52,5 @@ test_that("R file processing works and creates vignettes", {
     recursive = TRUE,
     force = TRUE
   )
+  remove.packages("rfiletest")
 })